add api get list mac + chỉnh sửa api insert/update

+++ src/main/java/com/example/employee_management/Dto/EmployeeDeviceDto.java
@@ -12,6 +12,6 @@ public class EmployeeDeviceDto {
     private String mac;
 
     @DbColumnMapper("STATUS")
-    private int status;
+    private Long status;
 
 }
